airflow/airflow_random_user_api/requirements.txt
awswrangler==3.11.0
boto3==1.37.26
flake8==7.2.0
isort==6.0.1
pandas==2.2.3
requests==2.32.3
python-dotenv==1.1.0
